renamed stub files and corrected config file.
