+++ languages/python/exercises/concept/little-sisters-essay/.meta/config.json
@@ -4,5 +4,9 @@
       "github_username": "kimolivia",
       "exercism_username": "olivia"
     }
-  ]
+  ],
+  "editor": {
+    "solution_files": ["string_methods.py"],
+    "test_files": ["string_methods_test.py"]
+  }
 }
